% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{constants}
\alias{constants}
\title{Calculate constants across estimation}
\usage{
constants(sign_level, reference = c("normal"))
}
\arguments{
\item{sign_level}{A numeric value between 0 and 1 that determines the cutoff
in the reference distribution against which observations are judged as
outliers or not.}

\item{reference}{A character vector of length 1 that denotes a valid
reference distribution.}
}
\value{
Returns a list that stores values that are constant across the
estimation to be accessed throughout the calculations.
The following elements are stored (in order): the significance level to
determine the cutoff value, the reference distribution as a character vector,
the numeric cutoff value to determine outliers, the bias correction factor
to account for the fact that even under the null hypothesis of no outliers,
there will be some false positives that incorrectly classified as outliers.
}
\description{
\code{constants} calculates various values that do not change across the
estimation and records them in a list.
}
\examples{
# constant values for a normal distribution with significance level 0.05
# observations with absolute residuals > 1.96 would be classified as outliers
cons <- constants(sign_level = 0.05, reference = "normal")
print(cons)
}
