% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{conv_diff}
\alias{conv_diff}
\title{L2 norm between two most recent estimates}
\usage{
conv_diff(current, counter)
}
\arguments{
\item{current}{A list object of class \code{"robust2sls"}.}

\item{counter}{An integer denoting the number of the current iteration.}
}
\value{
\code{conv_diff} returns a numeric value, which is the L2 norm
of the difference between the most recent and the previous parameter
estimates. The L2 norm is the sum of squared differences of the estimates.
}
\description{
\code{conv_diff} uses an object of class \code{"robust2sls"} to calculate the
L2 norm (sum of squared differences) between the most recent outlier-robust
iteration and the previous iteration estimates.
}
\keyword{internal}
