% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{estimate_param}
\alias{estimate_param}
\title{Estimation of moments of the data}
\usage{
estimate_param(robust2SLS_object, iteration)
}
\arguments{
\item{robust2SLS_object}{An object of class \code{"robust2sls"} for which
the moments will be calculated.}

\item{iteration}{An integer >= 0 specifying based on which model iteration
the moments should be estimated. The model iteration affects which
observations are determined to be outliers and these observations will hence
be excluded during the estimation of the moments.}
}
\value{
\code{estimate_param} returns a list with a similar structure as the
output of the Monte Carlo functionality \link{generate_param}. Hence, the
resulting list can be given to the function \link{gauge_avar} as argument
\code{parameters} to return an estimate of the asymptotic variance of the
gauge.
}
\description{
DO NOT USE YET!
\code{estimate_param} can be used to estimate certain moments of the data
that are required for calculating the asymptotic variance of the gauge. Such
moments are the covariance between the standardised first stage errors and
the structural error \eqn{\Omega}, the covariance matrix of the first stage
errors \eqn{\Sigma}, the first stage parameter matrix \eqn{\Pi}, and more.
}
\section{Warning}{

The function is not yet fully developed. The estimators of the moments are
at the moment not guaranteed to be consistent for the population moments. DO
NOT USE!
}

