% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{evaluate_boot}
\alias{evaluate_boot}
\title{Evaluate bootstrap results}
\usage{
evaluate_boot(r2sls_boot, iterations)
}
\arguments{
\item{r2sls_boot}{An object of \link{class} \code{"r2sls_boot"}, as returned
by \link{case_resampling}.}

\item{iterations}{An integer or numeric vector with values >= 0 specifying
which bootstrap results to evaluate.}
}
\value{
\code{evaluate_boot} returns a data frame with the bootstrap and the
  theoretical standard errors. Each row corresponds to a different iteration
  step while each column refers to the parameters whose standard errors are
  produced.
}
\description{
Evaluate bootstrap results
}
