% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{extract_boot}
\alias{extract_boot}
\title{Extracts bootstrap results for a specific iteration}
\usage{
extract_boot(r2sls_boot, iteration)
}
\arguments{
\item{r2sls_boot}{An object of \link{class} \code{"r2sls_boot"}, as returned
by \link{case_resampling}.}

\item{iteration}{An integer >= 0 specifying which bootstrap results to
extract.}
}
\value{
\code{extract_boot} returns a matrix with the bootstrap results for
a specific iteration.#'
}
\description{
Extracts bootstrap results for a specific iteration
}
