% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{generate_data}
\alias{generate_data}
\title{Random data of 2SLS model (Monte Carlo)}
\usage{
generate_data(parameters, n)
}
\arguments{
\item{parameters}{A list with 2SLS model parameters as created by
\link{generate_param}.}

\item{n}{Sample size to be drawn.}
}
\value{
\code{generate_data} returns a data frame with \code{n} rows
(observations) and the following variables of the 2SLS model: dependent
variable y, exogenous regressors x1, endogenous regressors x2, structural
error u, outside instruments z2, first stage projection errors r1 (identical
to zero) and r2.
}
\description{
\code{generate_data} draws random data for a 2SLS model given the parameters.
}
