% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monte_carlo.R
\name{mc}
\alias{mc}
\title{Monte Carlo simulations}
\usage{
mc(
  M,
  n,
  seed,
  parameters,
  formula,
  ref_dist,
  sign_level,
  initial_est,
  iterations,
  shuffle = FALSE,
  shuffle_seed,
  split = 0.5
)
}
\arguments{
\item{M}{Number of replications.}

\item{n}{Sample size for each replication.}

\item{seed}{Random seed for the iterations.}

\item{parameters}{A list as created by \link{generate_param} that specifies
the true model.}

\item{formula}{A formula that specifies the 2SLS model to be estimated. The
format has to follow \code{y ~ x1 + x2 | x1 + z2}, where \code{y} is the
dependent variable, \code{x1} are the exogenous regressors, \code{x2} the
endogenous regressors, and \code{z2} the outside instruments.}

\item{ref_dist}{A character vector that specifies the reference distribution
against which observations are classified as outliers. \code{"normal"} refers
to the normal distribution.}

\item{sign_level}{A numeric value between 0 and 1 that determines the cutoff
in the reference distribution against which observations are judged as
outliers or not.}

\item{initial_est}{A character vector that specifies the initial estimator
for the outlier detection algorithm. \code{"robustified"} means that the full
sample 2SLS is used as initial estimator. \code{"saturated"} splits the
sample into two parts and estimates a 2SLS on each subsample. The
coefficients of one subsample are used to calculate residuals and determine
outliers in the other subsample. \code{"user"} allows the user to specify a
model based on which observations are classified as outliers.}

\item{iterations}{An integer >= 0 that specifies how often the outlier
detection algorithm is iterated and for which summary statistics will be
calculated. The value \code{0} means that outlier classification based on the
initial estimator is done.}

\item{shuffle}{A logical value or \code{NULL}. Only used if
\code{initial_est == "saturated"}. If \code{TRUE} then the sample is shuffled
before creating the subsamples.}

\item{shuffle_seed}{An integer value that will set the seed for shuffling the
sample or \code{NULL}. Only used if \code{initial_est == "saturated"} and
\code{shuffle == TRUE}.}

\item{split}{A numeric value strictly between 0 and 1 that determines
in which proportions the sample will be split.}
}
\description{
NOTE: \code{mc} is currently not used. See \code{mc_grid} instead.
}
\details{
\code{mc} runs Monte Carlo simulations for a set of known population
parameters and a certain specification of the outlier detection algorithm.
}
