% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_r2sls}
\alias{new_r2sls}
\title{Constructor of r2sls class}
\usage{
new_r2sls(x = list())
}
\arguments{
\item{x}{A list with components of the "\code{r2sls}" class.}
}
\description{
\code{new_r2sls} turns a list into an object of \link{class} "\code{r2sls}"
}
\section{Warning}{

Only checks that the input is a list but not that its components match the
requirements of the "\code{r2sls}" class. Use the validator function
\code{validate_r2sls} for that purpose.
}

\examples{
\dontrun{
new_r2sls(1:10)
}
}
