% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{new_robust2sls}
\alias{new_robust2sls}
\title{Constructor of robust2sls class}
\usage{
new_robust2sls(x = list())
}
\arguments{
\item{x}{A list with components of the "\code{robust2sls}" class.}
}
\description{
\code{new_robust2sls} turns a list into an object of \link{class}
"\code{robust2sls}"
}
\section{Warning}{

Only checks that the input is a list but not that its components match the
requirements of the "\code{robust2sls}" class. Use the validator function
\code{validate_robust2sls} for that purpose.
}

\examples{
\dontrun{
new_robust2sls(1:10)
}
}
