% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{outlier}
\alias{outlier}
\title{Outlier history of single observation}
\usage{
outlier(robust2sls_object, obs)
}
\arguments{
\item{robust2sls_object}{An object of class \code{"robust2sls"}.}

\item{obs}{An index (row number) of an observation}
}
\description{
\code{outlier} takes a \code{"robust2sls"} object and the index of a specific
observation and returns its history of classification across the different
iterations contained in the \code{"robust2sls"} object.
}
