% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{outliers}
\alias{outliers}
\title{Number of outliers}
\usage{
outliers(robust2sls_object, iteration)
}
\arguments{
\item{robust2sls_object}{An object of class \code{"robust2sls"}.}

\item{iteration}{An integer >= 0 representing the iteration for which the
outliers are calculated.}
}
\value{
\code{outliers} returns the number of outliers  for a given iteration
as determined by the outlier-detection algorithm.
}
\description{
\code{outliers} calculates the number of outliers from a \code{"robust2sls"}
object for a given iteration.
}
