% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{plot.robust2sls}
\alias{plot.robust2sls}
\title{Plotting of standardised residuals and outliers}
\usage{
\method{plot}{robust2sls}(x, iteration = NULL, ...)
}
\arguments{
\item{x}{An object of \link{class} \code{"robust2sls"}.}

\item{iteration}{Either \code{NULL} (default) or an integer specifying the
iteration that should be plotted. The default uses the final model.}

\item{...}{Arguments to be passed to methods, see \link[base]{plot}.}
}
\value{
\code{plot.robust2sls} returns a graph of \link{class}
\link[ggplot2]{ggplot}.
}
\description{
Plot method for objects of \link{class} \code{"robust2sls"}. Plots the
standardised residuals of non-missing observations for a given iteration of
the outlier-detection algorithm and distinguishes whether an observation is
classified as an outlier by colour.
}
