% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{print.robust2sls}
\alias{print.robust2sls}
\title{Helper of robust2sls class}
\usage{
\method{print}{robust2sls}(x, verbose = FALSE, ...)
}
\arguments{
\item{x}{An object of \link{class} \code{"robust2sls"}.}

\item{verbose}{A logical value, \code{TRUE} or \code{FALSE}, determining
whether detailed (\code{TRUE}) or shortened (\code{FALSE}) should be printed.}

\item{...}{Further arguments passed to or from other methods, see
\link[base]{print}.}
}
\value{
No return value, prints model summary.
}
\description{
\code{robust2sls} allows the user to create an object of \link{class}
\code{"robust2sls"} by specifying the different components of the list. The
validator function \code{validate_robust2sls} is called at the end to ensure
that the resulting object is a valid object of \link{class}
\code{"robust2sls"}.
}
\details{
Printing summary output

Print method for objects of \link{class} \code{"robust2sls"}. Prints a
high-level summary of the settings and results of the outlier-detection
algorithm.
}
