% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{r2sls}
\alias{r2sls}
\title{Helper of r2sls class}
\usage{
r2sls(data, coefficients, gamma, iterations, converged)
}
\arguments{
\item{data}{A dataframe or matrix containing the original data used in the
estimation.}

\item{coefficients}{A numeric vector of the second stage coefficients.}

\item{gamma}{A numeric scalar between 0 and 1 representing the tail quantiles
that determine the cutoff value for judging outliers.}

\item{iterations}{Determines the number of iterations of the outlier
detection algorithm. Either an integer >= 0 or a character \code{"inf"}
representing iteration until convergence.}

\item{converged}{A logical value whether the algorithm has converged.}
}
\description{
\code{r2sls} allows the user to create an object of \link{class}
"\code{r2sls}" by specifying the different components of the list. The
validator function \code{validate_r2sls} is called at the end to ensure that
the resulting object is a valid object of \link{class} "\code{r2sls}".
}
