% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_estimators.R
\name{robustified_init}
\alias{robustified_init}
\title{Robustified 2SLS (full sample initial estimator)}
\usage{
robustified_init(data, formula, cutoff)
}
\arguments{
\item{data}{A dataframe or matrix containing the data used in the estimation.}

\item{formula}{A formula in the format \code{y ~ x1 + x2 | x2 + z2} where
\code{y} is the dependent variable, \code{x1} are the exogenous regressors,
\code{x2} the endogenous regressors, and \code{z2} the outside instruments.}

\item{cutoff}{A numeric cutoff value used to judge whether an observation
is an outlier or not. If its absolute value is larger than the cutoff value,
the observations is classified as being an outlier.}
}
\description{
\code{robustified_init} estimates the full sample 2SLS model, which is used
as the initial estimator for the iterative procedure.
}
