% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_estimators.R
\name{saturated_init}
\alias{saturated_init}
\title{Saturated 2SLS (split-sample initial estimator)}
\usage{
saturated_init(data, formula, cutoff, shuffle, shuffle_seed, split = 0.5)
}
\arguments{
\item{data}{A dataframe or matrix containing the data used in the estimation.}

\item{formula}{A formula in the format \code{y ~ x1 + x2 | x2 + z2} where
\code{y} is the dependent variable, \code{x1} are the exogenous regressors,
\code{x2} the endogenous regressors, and \code{z2} the outside instruments.}

\item{cutoff}{A numeric cutoff value used to judge whether an observation
is an outlier or not. If its absolute value is larger than the cutoff value,
the observations is classified as being an outlier.}

\item{shuffle}{A logical value (TRUE or FALSE) whether the sample should be
split into sub-samples randomly. If FALSE, the sample is simply cut into two
parts using the original order of the supplied data set.}

\item{shuffle_seed}{A numeric value that sets the seed for shuffling the
data set before splitting it. Only used if \code{shuffle == TRUE}.}

\item{split}{A numeric value strictly between 0 and 1 that determines
in which proportions the sample will be split.}
}
\description{
\code{saturated_init} splits the sample into two sub-samples. The 2SLS model
is estimated on both sub-samples and the estimates of one sub-sample are
used to calculate the residuals and hence outliers from the other sub-sample.
}
\section{Warning}{

The estimator may have bad properties if the \code{split} is too unequal and
the sample size is not large enough.
}

