% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_estimators.R
\name{user_init}
\alias{user_init}
\title{User-specified initial estimator}
\usage{
user_init(data, formula, cutoff, user_model)
}
\arguments{
\item{data}{A dataframe.}

\item{formula}{A formula in the format \code{y ~ x1 + x2 | x2 + z2} where
\code{y} is the dependent variable, \code{x1} are the exogenous regressors,
\code{x2} the endogenous regressors, and \code{z2} the outside instruments.}

\item{cutoff}{A numeric cutoff value used to judge whether an observation
is an outlier or not. If its absolute value is larger than the cutoff value,
the observations is classified as an outlier.}

\item{user_model}{A model object of \link{class} \link[AER]{ivreg} whose
parameters are used to calculate the residuals.}
}
\description{
\code{user_init} uses a model supplied by the user as the initial estimator.
Based on this estimator, observations are classified as outliers or not.
}
\section{Warning}{

Check REFERENCE TO PAPER about conditions on the initial estimator that
should be satisfied for the initial estimator (e.g. they have to be Op(1)).
}

