% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{validate_r2sls}
\alias{validate_r2sls}
\title{Validator of r2sls class}
\usage{
validate_r2sls(x)
}
\arguments{
\item{x}{An object whose validity of class "\code{r2sls}" is tested.}
}
\description{
\code{validate_2rsls} checks that the input is a valid object of \link{class}
"\code{r2sls}"
}
