% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{validate_robust2sls}
\alias{validate_robust2sls}
\title{Validator of robust2sls class}
\usage{
validate_robust2sls(x)
}
\arguments{
\item{x}{An object whose validity of class \code{"robust2sls"} is tested.}
}
\value{
If the object is a valid \code{"robust2sls"} object then the function
returns the object. No return value otherwise.
}
\description{
\code{validate_robust2sls} checks that the input is a valid object of
\link{class} \code{"robust2sls"}.
}
