% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gauge.R
\name{outlier}
\alias{outlier}
\title{Outlier history of single observation}
\usage{
outlier(robust2sls_object, obs)
}
\arguments{
\item{robust2sls_object}{An object of class \code{"robust2sls"}.}

\item{obs}{An index (row number) of an observation}
}
\value{
\code{outlier} returns a vector that contains the 'type' value for
the given observations across the different iterations. There are three
possible values: 1 if the observations is judged to be an outlier, 0 if not,
and -1 if any of its x, y, or z values required for estimation is missing.
}
\description{
\code{outlier} takes a \code{"robust2sls"} object and the index of a specific
observation and returns its history of classification across the different
iterations contained in the \code{"robust2sls"} object.
}
